# Improved BLAST performance with multithreading and JSON output type