Performance changes on 2kb query:

(Before)
infile generation: 0.003701925277709961 seconds
blast command generation: 0.0006384849548339844 seconds
**blast cl call: 375.5496497154236 seconds**
opening outfile: 0.0005927085876464844 seconds
**reading outfile into NCBIXML: 74.45608043670654 seconds**
result reading: 7.70909857749939 seconds
**_Single-threaded XML time: 458.09611415863037 seconds_**

(After)
infile generation: 0.0029015541076660156 seconds
blast command generation: 0.0012469291687011719 seconds
**blast cl call: 131.6555197238922 seconds**
opening outfile: 0.0009796619415283203 seconds
**reading outfile into json: 2.003464698791504 seconds**
result reading: 5.763728380203247 seconds
**_4-threaded JSON time: 139.61657047271729 seconds_**