@benjiec 2022-08-08
Feel free to merge afterward also may want to check with Jon or DevOps how big our machine is for Edge Rancher, and if we can make it bigger so we can run more than 4 threads?